% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena.R
\name{arena_new}
\alias{arena_new}
\title{Creates arena object}
\usage{
arena_new(
  live = FALSE,
  N = 500,
  fi_N = NULL,
  fi_B = 10,
  grid_points = 101,
  shap_B = 10,
  mc.cores = 1
)
}
\arguments{
\item{live}{Defines if arena should start live server or generate static json}

\item{N}{number of observations used to calculate dependence profiles}

\item{fi_N}{number of observations used in feature importance}

\item{fi_B}{Number of permutation rounds to perform each variable
in feature importance}

\item{grid_points}{number of points for profile}

\item{shap_B}{Numer of random paths in SHAP}

\item{mc.cores}{Number of cores used in calculating static arena}
}
\value{
Empty \code{arena_static} of \code{arena_live} class object
}
\description{
Creates arena object
}
