% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena.R
\name{arena_push_model}
\alias{arena_push_model}
\title{Adds model to arena}
\usage{
arena_push_model(arena, explainer)
}
\arguments{
\item{arena}{live or static arena object}

\item{explainer}{Explainer created using \code{DALEX::explain}}
}
\value{
Updated arena object
}
\description{
If arena is static it will start calculations for all already pushed
observations and global plots. If arena is live, then plots will be
calculated on demand, after calling \code{arena_run}.
}
