% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena.R
\name{arena_push_observations}
\alias{arena_push_observations}
\title{Adds new observations to arena}
\usage{
arena_push_observations(arena, observations)
}
\arguments{
\item{arena}{live or static areana object}

\item{observations}{dataframe of new observations}
}
\value{
Updated arena object
}
\description{
If arena is static it will start calculations for all already pushed
models. If arena is live, then plots will be calculated on demand,
after calling \code{arena_run}.
}
