% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena.R
\name{arena_upload}
\alias{arena_upload}
\title{Upload generated json file from static arena}
\usage{
arena_upload(arena, open_browser = TRUE, append_data = FALSE)
}
\arguments{
\item{arena}{Static arena object}

\item{open_browser}{Whether to open browser with new session}

\item{append_data}{Whether to append data to already existing session}
}
\value{
not modified arena object
}
\description{
By default function opens browser with new arena session. Appending data to
already existing session is also possible usign argument \code{append_data}
}
