% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_accumulated_dependence}
\alias{get_accumulated_dependence}
\title{Internal function for calculating Accumulated Dependence}
\usage{
get_accumulated_dependence(explainer, variable, params)
}
\arguments{
\item{explainer}{Explainer created using \code{DALEX::explain}}

\item{variable}{Name of variable}

\item{params}{Params from arena object}
}
\value{
Plot data in Arena's format
}
\description{
Internal function for calculating Accumulated Dependence
}
