% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_attributes}
\alias{get_attributes}
\title{Returns attributes for all params}
\usage{
get_attributes(arena, param_type = NULL)
}
\arguments{
\item{arena}{live or static arena object}

\item{param_type}{Type of param. One of
\itemize{
  \item{model}
  \item{variable}
  \item{dataset}
  \item{observation}
}}
}
\value{
List of attributes or named list of lists of attributes for each param type.
}
\description{
When \code{param_type} is not NULL, then function returns list of objects.
Each object represents one of available attribute for specified param type.
Field \code{name} is attribute name and field \code{values} is mapped list
of available params to list of value of this attribute for that param.
When \code{param_type} is NULL, then function returns list with keys for
each param type and values are lists described above.
}
