% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_break_down}
\alias{get_break_down}
\title{Internal function for calculating Break Down}
\usage{
get_break_down(explainer, observation, params)
}
\arguments{
\item{explainer}{Explainer created using \code{DALEX::explain}}

\item{observation}{One row data frame observation}

\item{params}{Params from arena object}
}
\value{
Plot data in Arena's format
}
\description{
Internal function for calculating Break Down
}
