% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_ceteris_paribus}
\alias{get_ceteris_paribus}
\title{Internal function for calculating Ceteris Paribus}
\usage{
get_ceteris_paribus(explainer, observation, variable, params)
}
\arguments{
\item{explainer}{Explainer created using \code{DALEX::explain}}

\item{observation}{One row data frame observation}

\item{variable}{Name of variable}

\item{params}{Params from arena object}
}
\value{
Plot data in Arena's format
}
\description{
Internal function for calculating Ceteris Paribus
}
