% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_dataset_attributes}
\alias{get_dataset_attributes}
\title{Generates list with attributes of a dataset}
\usage{
get_dataset_attributes(arena, dataset)
}
\arguments{
\item{arena}{live or static arena object}

\item{dataset}{List with following elements 
\itemize{
  \item{dataset}{ Data frame}
  \item{target}{ Name of one column from data frame that is used as target variable}
  \item{label}{ Label for dataset to be displayed in Arena}
  \item{variables}{ vector of column names from data frame without target}
}}
}
\value{
simple list with attributes of given dataset
}
\description{
Generates list with attributes of a dataset
}
