% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_dataset_plots}
\alias{get_dataset_plots}
\title{Internal function for calculating exploratory data anaylysis plots}
\usage{
get_dataset_plots(dataset, params)
}
\arguments{
\item{dataset}{List with following elements 
\itemize{
  \item{dataset}{ Data frame}
  \item{target}{ Name of one column from data frame that is used as target variable}
  \item{label}{ Label for dataset to be displayed in Arena}
  \item{variables}{ vector of column names from data frame without target}
}}

\item{params}{Params from arena object}
}
\value{
list of generated plots' data
}
\description{
Function runs all plot generating methods for given dataset
}
