% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{get_datasets_list}
\alias{get_datasets_list}
\title{Generates list of datasets' labels}
\usage{
get_datasets_list(arena)
}
\arguments{
\item{arena}{live or static arena object}
}
\value{
list of datasets' labels
}
\description{
Generates list of datasets' labels
}
