% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_feature_importance}
\alias{get_feature_importance}
\title{Internal function for calculating feature importance}
\usage{
get_feature_importance(explainer, vars, params)
}
\arguments{
\item{explainer}{Explainer created using \code{DALEX::explain}}

\item{vars}{Variables names for which feature importance should be calculated}

\item{params}{Params from arena object}
}
\description{
Internal function for calculating feature importance
}
