% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_funnel_measure}
\alias{get_funnel_measure}
\title{Internal function for calculating funnel measure}
\usage{
get_funnel_measure(explainer, params)
}
\arguments{
\item{explainer}{Explainer created using \code{DALEX::explain}}

\item{params}{Params from arena object}
}
\value{
Plot data in Arena's format
}
\description{
Internal function for calculating funnel measure
}
