% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_global_plots}
\alias{get_global_plots}
\title{Internal function for calculating global plots}
\usage{
get_global_plots(explainer, params)
}
\arguments{
\item{explainer}{Explainer created using \code{DALEX::explain}}

\item{params}{Params from arena object}
}
\value{
list of generated plots' data
}
\description{
Function runs all plot generating methods for given explainer
}
