% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{get_json_structure}
\alias{get_json_structure}
\title{Prepare object ready to change into json}
\usage{
get_json_structure(arena)
}
\arguments{
\item{arena}{live or static arena object}
}
\description{
Prepare object ready to change into json
}
