% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_message_output}
\alias{get_message_output}
\title{Internal function for returning message as plot data}
\usage{
get_message_output(output, type, msg)
}
\arguments{
\item{output}{existing plot data to be overwritten}

\item{type}{type of message "info" or "error"}

\item{msg}{message to be displayed}
}
\value{
Plot data in Arena's format
}
\description{
This method modify exisiting plot's data in Arena's format
to show message instead of chart.
}
