% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_metrics}
\alias{get_metrics}
\title{Internal function for calculating model performance metrics}
\usage{
get_metrics(explainer, params)
}
\arguments{
\item{explainer}{Explainer created using \code{DALEX::explain}}

\item{params}{Params from arena object}
}
\value{
Plot data in Arena's format
}
\description{
Internal function for calculating model performance metrics
}
