% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_model_attributes}
\alias{get_model_attributes}
\title{Generates list with attributes of a model}
\usage{
get_model_attributes(arena, explainer)
}
\arguments{
\item{arena}{live or static arena object}

\item{explainer}{Explainer created using \code{DALEX::explain}}
}
\value{
simple list with attributes of given model
}
\description{
Generates list with attributes of a model
}
