% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_observation_attributes}
\alias{get_observation_attributes}
\title{Generates list with attributes of an observation}
\usage{
get_observation_attributes(arena, observation)
}
\arguments{
\item{arena}{live or static arena object}

\item{observation}{One row data frame observation}
}
\value{
simple list with attributes of given observation
}
\description{
Generates list with attributes of an observation
}
