% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_rec}
\alias{get_rec}
\title{Internal function for calculating regression error characteristic}
\usage{
get_rec(explainer, params)
}
\arguments{
\item{explainer}{Explainer created using \code{DALEX::explain}}

\item{params}{Params from arena object}
}
\value{
Plot data in Arena's format
}
\description{
Internal function for calculating regression error characteristic
}
