% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_roc}
\alias{get_roc}
\title{Internal function for calculating receiver operating curve}
\usage{
get_roc(explainer, params)
}
\arguments{
\item{explainer}{Explainer created using \code{DALEX::explain}}

\item{params}{Params from arena object}
}
\value{
Plot data in Arena's format
}
\description{
Internal function for calculating receiver operating curve
}
