% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_shap_values}
\alias{get_shap_values}
\title{Internal function for calculating Shapley Values}
\usage{
get_shap_values(explainer, observation, params)
}
\arguments{
\item{explainer}{Explainer created using \code{DALEX::explain}}

\item{observation}{One row data frame observation to calculate Shapley Values}

\item{params}{Params from arena object}
}
\value{
Plot data in Arena's format
}
\description{
Internal function for calculating Shapley Values
}
