% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_variable_against_another}
\alias{get_variable_against_another}
\title{Internal function for variable against another plot}
\usage{
get_variable_against_another(dataset, variable, params)
}
\arguments{
\item{dataset}{List with following elements 
\itemize{
  \item{dataset}{ Data frame}
  \item{target}{ Name of one column from data frame that is used as target variable}
  \item{label}{ Label for dataset to be displayed in Arena}
  \item{variables}{ vector of column names from data frame without target}
}}

\item{variable}{Name of primary variable}

\item{params}{Params from arena object}
}
\value{
Plot data in Arena's format
}
\description{
Internal function for variable against another plot
}
