% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_variable_attributes}
\alias{get_variable_attributes}
\title{Generates list with attributes of an variable}
\usage{
get_variable_attributes(arena, variable)
}
\arguments{
\item{arena}{live or static arena object}

\item{variable}{Name of variable}
}
\value{
simple list with attributes of given variable
}
\description{
Generates list with attributes of an variable
}
