% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{get_variables_list}
\alias{get_variables_list}
\title{Generates list of unique variables(without y) from each explainer}
\usage{
get_variables_list(arena)
}
\arguments{
\item{arena}{live or static arena object}
}
\description{
Generates list of unique variables(without y) from each explainer
}
