% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena.R
\name{save_arena}
\alias{save_arena}
\title{Save generated json file from static arena}
\usage{
save_arena(arena, filename = "data.json", pretty = FALSE)
}
\arguments{
\item{arena}{Static arena object}

\item{filename}{Name of output file}

\item{pretty}{whether to generate pretty and easier to debug JSON}
}
\value{
not modified arena object
}
\description{
Save generated json file from static arena
}
