% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{split_multiclass_explainer}
\alias{split_multiclass_explainer}
\title{Splits multiclass explainer into multiple classification explainers}
\usage{
split_multiclass_explainer(explainer)
}
\arguments{
\item{explainer}{Multiclass explainer created using \code{DALEX::explain}}
}
\value{
list of explainers
}
\description{
Splits multiclass explainer into multiple classification explainers
}
