% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{truncate_vector}
\alias{truncate_vector}
\title{Internal function for pretty truncationg params list}
\usage{
truncate_vector(vec, size = 6)
}
\arguments{
\item{vec}{vector to be truncated}

\item{size}{elements with index greater than size will be truncated}
}
\value{
string with collapsed and truncated input vector
}
\description{
Internal function for pretty truncationg params list
}
