% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena.R
\name{upload_arena}
\alias{upload_arena}
\title{Upload generated json file from static arena}
\usage{
upload_arena(
  arena,
  open_browser = TRUE,
  append_data = FALSE,
  arena_url = "https://arena.drwhy.ai/",
  pretty = FALSE
)
}
\arguments{
\item{arena}{Static arena object}

\item{open_browser}{Whether to open browser with new session}

\item{append_data}{Whether to append data to already existing session}

\item{arena_url}{URL of Arena dashboard instance}

\item{pretty}{whether to generate pretty and easier to debug JSON}
}
\value{
not modified arena object
}
\description{
By default function opens browser with new arena session. Appending data to
already existing session is also possible using argument \code{append_data}
}
