% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{validate_new_dataset}
\alias{validate_new_dataset}
\title{Checks if it is safe do add new dataset to the arena object}
\usage{
validate_new_dataset(arena, dataset, target, label)
}
\arguments{
\item{arena}{live or static arena object}

\item{dataset}{data frame for data analysis}

\item{target}{name of target variable}

\item{label}{name of dataset}
}
\value{
None
}
\description{
Checks if it is safe do add new dataset to the arena object
}
