% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{validate_new_model}
\alias{validate_new_model}
\title{Checks if it is safe do add new model to arena object}
\usage{
validate_new_model(arena, explainer)
}
\arguments{
\item{arena}{live or static arena object}

\item{explainer}{Explainer created usign \code{DALEX::explain}}
}
\description{
Function checks if label is not already used
}
