% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{validate_new_observations}
\alias{validate_new_observations}
\title{Checks if it is safe do add new observations to the arena object}
\usage{
validate_new_observations(arena, observations)
}
\arguments{
\item{arena}{live or static arena object}

\item{observations}{data frame of new observations}
}
\value{
None
}
\description{
Function checks if rownames are not already used and call stop if
there is at least one conflict.
}
