% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{get_json_structure}
\alias{get_json_structure}
\title{Prepare object ready to change into json}
\usage{
get_json_structure(arena)
}
\arguments{
\item{arena}{live or static arena object}
}
\value{
Object for direct conversion into json
}
\description{
Function converts object with class \code{arena_live} or \code{arena_static}
to object with structure accepted by Arena. See \href{https://github.com/ModelOriented/Arena/tree/master/src/store/schemas}{list of schemas}.
}
