% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{get_observations_list}
\alias{get_observations_list}
\title{Generates list of rownames of each observation from each batch}
\usage{
get_observations_list(arena)
}
\arguments{
\item{arena}{live or static arena object}
}
\description{
Generates list of rownames of each observation from each batch
}
