% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena.R
\name{print.arena_live}
\alias{print.arena_live}
\title{Prints live arena summary}
\usage{
\method{print}{arena_live}(arena)
}
\arguments{
\item{arena}{\code{arena_live} object}
}
\description{
Prints live arena summary
}
