% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena.R
\name{print.arena_static}
\alias{print.arena_static}
\title{Prints static arena summary}
\usage{
\method{print}{arena_static}(x, ...)
}
\arguments{
\item{x}{\code{arena_static} object}

\item{...}{other parameters}
}
\description{
Prints static arena summary
}
