% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{validate_new_model}
\alias{validate_new_model}
\title{Checks if it is safe do add a new model to the arena object}
\usage{
validate_new_model(arena, explainer)
}
\arguments{
\item{arena}{live or static arena object}

\item{explainer}{Explainer created using \code{DALEX::explain}}
}
\description{
Function checks if explainer's label is not already used
}
