% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{validate_new_observations}
\alias{validate_new_observations}
\title{Checks if it is safe do add new observations to the arena object}
\usage{
validate_new_observations(arena, observations)
}
\arguments{
\item{arena}{live or static arena object}

\item{observations}{data frame of new observations}
}
\description{
Function checks if rowname of each row is not already used
}
